Add initial GPG sign support
It doesn't seem to quite work. It looks good in the SENT folder, but
it's been changed on the receiving side to no longer be
multipart/signed.

But even without signatures (defaulting to off because it doesn't
quite work yet) this changes all outgoing messages to be multipart
messages. That's needed for normal attachments, too.